requirements.txt
fastapi==0.115.13
pytest==8.3.5
python-dotenv==1.1.0
SQLAlchemy==2.0.41
sqlmodel==0.0.24
